Added new CL option '-b' to specify border width between cells
The `-b` option, or its long counterpart `--border-width`, expects an integer argument to set as the gap between two cells on display.
This allows to easily count the size of a contiguous row of cells in the same state: without borders, it becomes difficult two distinguish a row of 4 live cells from a row of 5.