Added function updateBoardWindow()
This function only modifies the *sel_x* and *sel_y* members of the structure, to refer to the current location (in cell-dimension precision) of the mouse pointer.

These members are then used in function renderBoardWindow() to highlight the hovered cell.

The type of the members has also change from `unsigned int` to `signed`, to correspond to `SDL_GetMouseState()`'s parameters type.