Took into account rules in board evolution.
Set the rules to the default value in `main()`.
Added static functions to determine, according to the current state of the cell, whether it will be alive on the next generation.
The functions are `willBeBorn` (called if the cell is currently dead) and `willSurvive`, called if the cell is alive.