Added dixtionary of names for common rules.
Reworked static function *validate_rules()* in `clop.c`: renamed to *rvalset()* and sets the rule variable itself instead of the `switch` in *getvals()*; uses function *getRuleFromName()* from `rules.h` module.
The latter is implemented using the binary search algorithm to look among an array of currently *32* names of CAs to find the provided one and return the correspondig rule -- returns `NULL` if it does not found the name.