Store the wrap flag in the board structure
The function member getCell is removed, and replaced with a bool member,
wrap. The macro getBoardCell is replaced with a function of identical
prototype; the function is implemented by using the new member to
delegate to either one of the internal getCell* functions.
The bool variable in updateBoard and the parameter to updateRow are
replaced with uses of this member.