Fix implementation of getBoardCell
The delegate functions returned a pointer to the cell, that could have
been NULL when the coordinates are invalid (no wrapping) the pointer was
not even dereferenced in getBoardCell.
The delegates now return the cell value directly, or false if the coords
are invalid.
Also, the forward declarations of the delegate functions are removed;
and the return value of the function is explicited in its documentation.