WIP: use a bit array to store the game board
The board source uses the bit array manipulation macros and function
to handle the game board bits instead of raw memcpys.
The internal function performing the update of each row of the board is
not yet updated, as it will be a more complex task.
The struct definition was already updated in the header by mistake
earlier, so it is not displayed here.