Remove function updateBoardWindow
The function was doing what the main loop was already able to do. The
variables local to main over_x and over_y are removed, and the struct
boardwindow members sel_x and sel_y are used instead; the update of the
selection coordinates in the mouse motion event handling block is
extracted from the if to be always run.