Do not use regex to check the rule string
The libc regex engine is not powerful enough to properly check the rule
string (lookaheads would be needed, but are not supported).
The regex compilation and usage are removed, and rvalset now parses the
string and checks the characters directly.

Also, the unneeded include of string.h is removed, and an include
comment is added for stdio.h.