Make the slash in rule string optional
The internal function rvalset is updated to allow the absence of the
slash character between the B part digits and the S character.
This is an extension from the original Golly format, provided for
convenience (although it does help visually, the slash really is not
necessary for the expression to be unambiguous).