Rename parameters to function getHoverCoord
The output parameters x and y are renamed i and j. These name are more
fitting, since in a mathematical context, these letters are used to
represent respectively column and row numbers in discrete 2-dimensional
objects (notably matrices), whereas x and y are usually real numbers
that represent coordinates in continuous 2-dimensional objects (planes).
The x and y are still used in the implementation of the function, as
variable that hold the screen coordinates of the mouse pointer as
returned by the SDL.