Implement function loadBoard
The function is implemented by first calling a helper function that
iterates the string board representation to count the number of
characters until the first new line (this gives the board width) and
then the number of newlines characters (this gives the board height).
The helper then delegates the board initialization ti initBoard, with as
dimensions the values computed beforehand.
Afterwards, the representation is iterated once again to determine the
state of individual cells.