Optimize loadBoard internal helper
The function iterates the board representation with a step of the width
of the board. This drastically cuts the looping, but will fail on an
ill-formed representation: if any row is not the proper length, if the
last row is shorter, the function will read outside the buffer. The
former case is checked, the latter cannot.
An error code in loadBoard is changed to not conflict with the new value
returned by the helper function.