Fix bugs in readFile implementation
The file cursor was left at the end of the file before reading, this
caused the next fread call to read nothing.
This error was dropped by the incorrect chaining of calls in the if
expression: the calls to fread and ferror were separated with a logical
and, but actually the logical or is the intended operation.

When overwriting the trailing newline, the index is now fixed; also the
terminator NUL is written on the last character allocated if not in
place of the trailing newline because apparently fread does not write
it.