Fix board state initialization in loadBoard
The looping over the string repr is fixed: the counter i is now only
used to represent the index in the cell array, and the char pointer
itself is incremented. This allows the two indices to evolve separately,
and this allows to skip the newline characters by incrementing once more
the char iterator but not the index.