Use r key to reset board
When the board has been initialized from a file, pressing r will reset
its state to the one initially read from the file (and, if the game is
running, pause it).

The variable repr in main is declared at function-level, and passed
through runApp to the event-handling function. It is freed at the end of
main along with the board and boardwindow objects.