Allow writing to standard output stream
The function writeFile is modified to interpret an input path of "-" as
an indication to write to stdout; the writing is implemented as a
delegating call from writeFile to a new internal function that is
basically a copy of writeFile without the fopen and fclose calls.

The new behavior is fully documented.