Allow reading from standard input stream
As for writeFile, readFile is modified to handle differently a path of
simply "-". It calls a new internal function that reads stdin in chunks
and appends them to a result string. The functions is implemented
differently from readFile because stdin is not a file and cannot be
fseek'd, this is explicited in a comment.

The new behavior is also documented.