Implement -i and -o options
The variable file in main and the parameter file of parseCommandLineArgs
are split in two, in_file and out_file. Each is assigned with the
relevant option argument in parseCommandLineArgs and passed to the
proper function calls in main.
Also, the call to isFile in main is removed, as the input file must
exist in order to read from it (it was optional when only the -f option
was implemented because it could mean only the output functionality was
desired).