Allow board initialization from RLE repr string
The RLE is a compressed board representation useful for huge patterns.
The format implemented here is the xlife one, with a header line
specifying the dimensions of the board and optionally the rule used.

Currently, only the parsing of the header line is implemented, with a
TODO tag in place of the future boazrd cells init code.