Fix several bugs in RLE loading
The row number of a run length was not given, making the internal
function setRunLength overwrite the first row every time it was called.

The repr string is no longer incremented a second time after it is set
to the character after the run length number, because the increment is
already done by the parsing loop in initCellsFromRLE.

In the latter function, the while loop is change to for so that the
increment of the string is done after the switch has been executed, not
before (this skipped the first character).

The boundary checking when incrementing the column number is changed
from greater-or-equal to strictly greater than the board width to allow
for the handling of the last column in each row (this is not applied to
the height checking to disallow a $ right before the terminal !.

The column number is reset to 0 when the row number is increased.