Fix incorrect behavior of parseCommandLineArgs
The function returned an error value when the number of options (with or
without arguments) and non-option arguments was different from the total
number of arguments. This is incorrect, since an option's argument can
be separated from its option by a space and counted as two arguments in
argc, while the counter was still incremented by only one.

The counter variable is removed altogether and this check is no longer
performed.