Improve handling of command-line options errors
The opttions string is prefixed with a colon, indicating to getopt to
return ':' for recognized options wih missing arguments. This
differentiates the case from an unrecognized option, where '?' is
returned. The variable opterr is set to 0, to indicate to getopt not to
print an error message. Cases for '?' and ':' are added to the options
switch block to handle the explicitly the error cases; '?' is not
treated as a fatal error since the execution may be able to proceed,
whereas an error coe is returned from the ':' label (a missing argument
should be reported immediately so the user can act quiclky). A message
is added to the default case, since now the code should not be able to
enter this label anymore.