Add a --help/--usage command line option
The option has no short form, because -h is already taken, and -u is not
assigned to keep the two options similar in behavior.
The usage message is added as an internal constant char array, and is
written in a new internal function to stderr. When either of the option
is met, the function parseCommandLineArgs returns a positive value. The
calling code is thus modified to exit with a success status when a
positive code is returned.