Add nmake-specific makefile
A file named Makefile is added back again, but this time uses
nmake-specific syntax. This is really sub-optimal, as the tool notably
lacks wilcard expansion and flexibility in its inference rules
definition. But it's the Windows sanctionned way, so if I'm going proper
Windows native that is an obligatory step.
Anyway, the Makefile iqs mostly ready, the named rules use the same
names as the GNU makefile. Only is missing the link to the SDL2 lib (a
placeholder TODO tag is present).
Most of the project structure is common to the two platforms, and a
substantial amount of configuration could be mutualized in a shared.mak
file that both makefiles include.
The program database file (the MS-specific catalog of debugging symbols)
is created in the root directory during compilation with debugging
enabled; all files ending in .pdb are added to the ignore list.