Link to SDL2
The library is merely added to the LDLIBS macro, as the linker expects
libraries to be passed as simple arguments like input object files
(contrary to ld which accepts them with the -l option). This expects the
lib file to be placed in a directory listed in ther env var LIB, in the
same way that #includes to SDL headers expects them to be in a directory
listed in the env var INCLUDE.
Commands invoking link in rules definitions are slightly changed, linker
options are passed before output option and input arguments.
The release flag is added to the linker flags when not in debug to
generate a production-ready executable (this may be a little
presomptuous, but it's also good future-proofing)