Make MSVC-friendly a function using restrict
The keyword is not recognized by MSVC, which is above all a C++
compiler, and C++ does not have the keyword. Fortunately, MSVC provides
its own equivalent as extension; which keyword is actually used is
abstracted away behind a macro valued according to the compiler's
predefined macros (only _WIN32 for now, will probably evolve when other
vendors become supported).