Implement getopt_long for non-GNU platforms
The function declared in a #ifdef block checking the GNU test macro is
defined at the bottom of the same source file, in another #ifndef block
checking the same macro.

The given implementation is minimal, the standard getopt behavior is not
strictly followed, and several GNU extensions are not supported; only
what was needed to the current program is implemented. Notably, setting
opterr to a non-zero value is a no-op: the flag is not used in the
implementation, and argv is not reordered, so the program will quit
processing the command line on the non-option argument, even if valid
options are present after it. Otherwise it is functional, and accepts
multiple short options in the same command-line token, short options
with attached argument, long options with argument attached (separated
with a = sign) or detached (on the next command-line token).
It also is certainly not optimal, and probably barely portable, but its
presence is already a portability fix.
Comments are added throughout the function and its internal helper
functions to explain more obscure parts.