Add an icon to the game window
The icon  pixel data is defined as a big array of unsigned shorts,
compiled into a SDL_Surface and assigned to the window.

Two data sets are given to the array, one corresponding to a 16x16 icon,
the other to a 64x64 one; the actual value being selected from the value
of the macro ICONSIZE (16 or 64).