Add a data directory with resources and icon files
The resources file (.rc) is minimal, it only designates the .ico file as
an icon, butthis is enough for it to be recognized.
The icon itself is the same as the image used as window icon, a black
and white down, left-oriented glider with cells separated by a thin grey
border, to mirror the aspect of the game grid, but with a transparent
background instead of white.

A Makefile variable is also defined to provide a name-independent handle
to the new directory.