Compile the resources file and link it to the exe
The file extensions rc and res are added to the list of suffixes nmake
should recognize as such.

An inference rule is added to compile any rc file in the data dir to a
res file in the obj dir (while this file is not an object file, it is
still an intermediate generated file, and in that this directory fits
most).
The link rule is modified: the res file is added as a dependency. When
the rule is invoked, the res file is then simply listed among the
files to link and the linker recognizes it.