Use snake_case everywhere
The mixed case styles hurt the readability and homogenity of the code
(snake case was used for structs and enums as per C convention, but
camel case was used for functions in an awkward attempt to shorten
their names, probably. All functions are renamed; most just have the
uppercase letters converted to their lowercase counterpart preceded
with an underscore, some names are also modified more deeply.