Remoe some compilation output
An echoed statement is removed in a test target rule.
Invocations of cl are silenced with @ because the command line is
starting to get very long and in clutters the screen, and the compiler
prints the name of the source file currently compiled anyway so this
information is not lost.