Rewrite grid init internal func for plain text
The function's loop no longer steps by width + 1, but uses strchr to
locate the next newline char and verify that the current line is the
intended width, or a comment.

This implementation is maybe slower than stepping fulle lines, but has
the advantage of being safe against accesses out of string bounds.
Also, we have to rely on libc having optimized their implementation of
strchr properly, so the overhead might be negligible. Anyway this
function is not supposed to be called very often.