Allow O in plain text formats instead of @
Both characters are recognized as meaning a live cell. The @ is kept
for personal preference and the O is added to conform to the common
usage. However, if both are met in the same file, a warning is printed.

The new behavior is documented in the READMEs. Also, a typo is fixed in
the French one.