Use the SDL's high-resolution timer feature
The function SDL_GetPerformanceTimer is used instead of SDL_GetTicks.
Since the granularity of the perf counter is not system-dependent
(whereas SDL_GetTicks is always microsecond precision), the hard-coded
1000 to convert the update rate to the frame duration is changed to a
call to SDL_GetPerformanceFrequency which gives the equivalent value.
Also, the macro SDL_TICKS_PASSED is no longer used since it is designed
for 32-bits integers only.