Handle CRLF line endings in input file
The function _init_grid_from_plain is modified to tolerate CR before
LF. This is implemented by assuming (with reason) that if the first
line ends with CRLF, all lines would. Then we check that all line
lengths including the CR are equal and once at the end decrement the
grid width to account for the CR to ignore.
The function load_grid then performs the input data checks and a branch
in the big if/else-if... block is added for CR, checking that it is
followed by a LF, processing it properly if it is, returning an error
code if it is not.

The input file is open in binary mode; this blocks Windows' fread to
preprocess the file data (notably converting CRLF to LF).
The trailing line break in the input data is checked for CRLF, not just
LF.