Simplify if expression in copy_bits
The expression used to check whether the source and destination bit
arrays have the same byte offset is changed to a simple comparison:
the bitwise AND on the difference of the two is not needed since both
alreeady are bitwise-AND-ed, and comparing the difference to 0 is
equivalent to comparing the values for equality.

Also, a comment is added to explain why the arrays are initially
modified by the eigth of their offset, and the comment explaining the
case where no common alignment is found is reworded.