Add a test source for bits.h functions and macros
The test source defines test functions for each macro or function
declared in the header, but not all are currently implemented.
Also defined is a master test function that calls the actual tests
functions.

The master function is declared as external function in the project's
main test source and invoked.

The bits.o object file is referenced in Makefile, in the list of
required objects for the tests compilation (and an obsolete object is
removed).