Fix copy_bits test function
The test cases loop was iterating up to a wrong value: the size of the
array of expected values was eight times the number of test iterations
(due to it being an array of pointers). The size of the array is thus
divided by the size of its elements in a typical XC fashion to finally
retrieve the proper number of test cases.