Return early if there is nothing left to copy...
... after the leading offset.

It can certainly happen that the full copy will take place within a
single byte, in which case the length correction statement will
underflow. A comparison is performed before updating the length, and if
it is less than or equal to the size of a byte minus the leading
offset, it means that the copy operation is finished and the function
returns.

Also, a typo is fixed in the length update statement comment.