Limit properly the copy of the partial 1st byte
In the implementation of copy_bits, the copy of the first byte was
performed up to 8 unconditionnally. Howoever, if the src_offset and the
length of the copy together do not reach a full byte, the copy would
overflow and unintendedly overwrite bits. This is fixed by making the
up end of the leading partial byte copy the minimum of 8 and the sum of
the source offset and the copy length.