Slightly improve copy_bits once again
The early return if the copy is finished after the partial leading byte
is copied is moved before the increments of the source and destination
pointers. The expression is also updated to be simpler and match the
definition of variable end, and a reference to this variable's comment
is added to the block's own comment.