Implementresizing the game window with the mouse
A case is added to the event-handling switch block for window events.
The block checks that the event is a resizing event, in which case it
computes the new grid dimensions and resizes the grid.

An internal function is added to transform a window dimension to a grid
one, the operation is basically an integer division of the window
dimension by the grid's cell size + border size.

The current resizing implementation is not ideal: the window can be
resized to sizes not exact multiple of the cell size, resulting in
thicker right and bottom borders.