Filter window resizing events
An event filter is set to limit the number of resizing events. In its
definition only the type "resized" of window events are checked (all
other types currently pass through unchanged), and on these the
dimension of the resize are changed to the corresponding grid
dimensions and, if the dimensions are identical to the current ones,
the event is blocked.
Changing the event data allows a simpler code in the event handling
block, merely passing the event data to the grid resize function.