Reorder if branches
A mostly negative if expression in the event filter function is changed
to the inverse expression and the branches are inverted, to improve
clarity.

Also, a comment is added to explain the early return, and a return
statement is removed to let the code fall through to the final return
statement.