Do not allow grid smaller than 3 cells in any dim
The if block checking the correctness of the new grid dimensions in the
event filter is added a new else-if branch, checking whether either
value is less than 3 and blocking the event in this case. Although it
won't stop from resizing the window to a smaller frame out of which the
grid would overflow, at least the grid itself will remain sane.