Pass user-supplied compilation flags to Make
In the definition of each of the four compilation/linkage variables
(CPPFLAGS, CFLAGS, LDFLAGS and LDLIBS), the variable itself is appended
so that values specified by the user on the command-line are passed
through to the compiler or linker instead of being simply overwritten.
To allow the variable to be referenced at the end of its redefinition,
temporary variables are introduced, holding the optimisation level and
debug flags.

This allows to pass arbitrary or teporary values without having to
modify the Makefile, but just by adding the variable to the environment
when invoking make, e.g.:

    CPPFLAGS=-DTEST_MACRO make test