Implement a minimum check for numerical opt args
The internal function _get_uint_value is modified to take another
argument, which is used as a lower bound for the numerical value for
the option.

Values passed are 3 for grid dimension options, 1 for the cell size and
update rate and 0 (check ignored since values are unsigned) for the
border width.