Refactor grid window init fucntion
The new macro is used instead of an explicit size conversion
expression. For this to work, assingments of the grid window members
corresponding to the cell size and border width to their respective
values are moved to the top of the function.
The grid member assingment is also moved, to respoect the order of the
function's parameters.