WIN32 Clamp window size properly after resize
After the window is resized with the mouse, its size is adjusted to fit
the grid nicely. Resizing events to the same grid size are no longer
filtered out so that they reach the event loop and the window is
properly clamped to the grid (cancelling the resize, but keeping a neat
window).