Only clamp the window to the grid on Windows
On X11, the call to SDL_SetWindowSize during handling of resize events
makes the window jolt around (sometimes disappearing from the monitor
frame) or grow vertically extremely fast (even if the resize is only
horizontal). This makes the app unusable, so the call is skipped when
not compiling for Windows (this is equivalent as compiling for X11 as
currently only these 2 window managers are supported).