Link to libX11 to enable incremental win resize
The library X11 is added to the list of libs linked against during
compilation in the GNU Makefile.
Window manager info is retrieved using the SDL. For this operation, the
SDL_VERSION statement is crucial--apparently the SDL function needs the
version to retrieve the proper window manager info.
This info is then used in calls to X11 functions that interact with the
window's hints, a set of values that dictate the window's behavior.
Here we set the width and height increment to the sum of the cell size
and border width, which corresponds to a unit in grid size.