Skip X11-related code if X11 is not available
The #includes and the X11-specific branch of the window manager are
wrapped in a #ifdef block, checking whether a SDL macro for X11 is
defined (I'm not sure that we are supposed to use this macro, but it is
far simpler this way).
The #endif is placed after the else keyword, but before the opening
bracket: this way, if the #ifdef block is skipped, the code is still a
valid anonymous block.