Implement properly the function compile_grid_rule
The function is implemented by replacing the grid field rule by an
array of bits where each bit is the next state of a cell whose current
neighbor configurations is the index in the array.
The field is cleared on grid initialization, so that during
"compilation" only bits for cell configs evolving into a live cell have
to be assigned.
The function is implemented lax in the rulestring format it accepts.
Ascending order and unicity within each of the B and S sections are
not checked. The "B" and "S" characters themselves are optional, and
indepentently of one another. However, either the "S" or the slash is
required to separate the parts. Also, any non-digit character generates
an error.
The internal functions to compute the next state from the rule string
are removed, and their invocations replaced with access to the new bit
array.
Also, the string duplication code in the load_grid internal helper is
removed, as mentioned that it would be.