Crudely allow + as a key's secondary sign
The keyboard event handling loop is modified to accept Shift+Equals as
an alias for the Plus key. This is explained in a comment; the reason
is that the Plus sign is on many keyboard layouts the secondary
character of the Equals key, accessed with the Shift modifier;
accepting Equals key presses with a Shift modifier is a crude (but
effective) way of handling these layouts.
Keyboard layouts where the Plus and Equals signs are not on the same
key usually have one where + is the principal character, pressing this
key is already handled in the event loop by the "case SDLK_PLUS:" line.
If then the Equals sign has its own key, pressing it with the Shift
modifier will still be handled. Although this quirk is not intended, it
is not really an inconvenient; it only gives an extra keyboard shortcut
"for free".