Change size (and meaning) of the grid rule array
The size is changed so that not only the number of live neighbors, but
rather the configuration of them is the bit index in the array.
Precisely, the three highest-order bits (the leftmost ones) correspond
to the respective states of three neighbors above, the fourth bit from
the left is the neighbor on the left, the fifth bit is the current
cell, the sixth bit is the neighbor on the right and the three
rightmost bits are the three neighbors on the row below.
While this will incur some redundancy in the states for isotropic and
(outer-)totalistic LLCAs, this will become crucial when I implement
anisotropic or non-totalistic LLCAs, where the positions of the live
neighbors matter as much as their number.

Obviously, the rule compiling code has now little left in common with
what it was before, obviously, and it does not actually work perfectly
yet--the tests fail for the moment. However, I chose to check in the
code, because the code is close from working as intended, and it is
already huge for a single commit.

The grid update code is also heavily modified to take into account this
new way of accessing the rule array; a bit set is made from the
neighbors' states and the current cell's.