Fix another rule compilation bug
This bug concerned the blocks for digits 5, 6, 7 and 8. It was due to
an improper bit operation, that reversed also the current cell bit,
while this bit should stay set to its original value. Instead, the
current cell bit is added to a bit mask of 111101111 and the bits
corresponding to the loop indices are unset.