Fix update_grid, #3
This fix concerns _update_cell once again. In the OR-combination of the
two state bytes, the second is cast to unsigned char. This is because
operands of the bitwise OR are promoted to int, and if the high order
bit is set, the signed char is interpreted as negative, and during int
promotion all higher bits will also be set, leading to a bit index far
too big (over 512 anyway).