Tweak rulestring compilation internal function
The function _handle_rule_digits now takes the rulestring pointer as
parameter, and successively checks the current character for digits 0
through 8, setting the corresponding bits in the rule buffer and
incrementing the pointer on match. This has two advantages: it
simplifies the code (somehow... At least by removing one indent level),
and it enforces that the digits be in ascending order in the
rulestring.
The return value is also changed: on OK, the modified pointer is
returned; if an invalid character (i.e. non-digit or not a slash,
capital S or NUL char) is encountered NULL is returned as error code.