Rename compile_grid_rule and move it
The function is made independent of the grid structure.
It is renamed compile_rulestring, and its grid parameter is changed to
a simple bytes array where the compilation result will be written;
also, the function's declaration is moved to rules.h since it is
conceptually more closely related to rules and rulestrings than to the
grid.
Include directives are added to the sources calling this function.