Introduce a cell state enumeration
The enum contains only two entries, DEAD and ALIVE. It is basically a
redefinition of a boolean type, but with more appropriate names for the
situation.
It is used as the return type of get_grid_cell and toggle_grid_cell and
the constants are used in some grid tests.