Handle system events earlier
The call to the system event handler is moved inside the event filter
function. This is because on Windows (at least) the filter is not run
on SDL's event pipeline, but on the system one, where it receives, in
real time, notably the window resizing events. (On the SDL event chain
the resizing events are all sent once the resizing is finished and the
incrementation is not observable.)