Support Windows window manager
When the code is compiled with the Windows window manager available
(i.e. on Windows), the runtime subsystem is checked against this window
manager's constant because it is now supported, but the code block is
empty (the resizing in increments is performed while handling the
system event), save for a comment explaining why so.