Fix the jittering when resizing the window
The call to resize_grid_window inside handle_system_event was
responsible for the window stuttering when resized. This call is simply
replaced by a modification of the event's size parameters to let the
window manager perform the resize, and a repaint call to update the
content of the new zone.