Specify a format for the output file
A parameter is added to run_app, giving a format for the output file.
It is passed through until get_grid_repr is called to generate the grid
representation to write to the file.
A default hard-coded value of plain text is used for this parameter
where the function is called.