Revert previous commit for non-Windows platforms
The code handling SDL_WINDOEVENT is reintroduced, wrapped in a #ifndef
block enabed only when not compiling with MSVC (for the time being,
compiling with MSVC is considered equivalent to being on Windows as
it is the only compiler I support for Windows).

The code reintroduced only resizes the grid if the window event is a
resize one, the grid window is not resized (since it is already at the
correct size due to the resize increment set earlier on the X11 win).