Rename project to Cyano
This name is inspired by the cyanobacteria, a family of bacteria known
to be among the first lifeforms to assemble in complex structures.
Also, it fits a personal convention of having project names start with
the first letter of the language it uses.

The name of the project is changed in the READMEs, the doxygen config
file, the platform-shared config Makefile and the Windows resources
declaration file.