License the project under CeCILL v2
This license was chosen for several reasons. First, it is written with
French law in mind, fixing some incompatibilities between it and the
GNU GPL (and other licenses with for example a "no warranty" clause,
for that matter). Moreover, it is a bilingual license, valid as much in
French and in English, and it fits well with the stated objective for
this project to be bilingual as well.