Prepare implementation of RLE write
The function get_grid_repr is refactored: the code that writes the
plain-text representation of the grid to the string to be returned is
extracted in a new internal function. Another equivalent function is
defined, intended to perform the same operation in RLE format. It is
currently empty, save for a NUL byte written to the first character of
the repr string, to properly terminate it.
The function among these two to which delegate is chosen in
get_grid_repr according to the value of the format parameter.