Trim unused memory for RLE repr
If the result RLE representation is lower than the maximum allocated
value (which will often be the case, that is the point of RLE after
all), the representation string is reallocated.
A NULL return from realloc is not explicitly handled, as the realloc is
the last operation performed before actually returning, so a NULL
return statement would be superfluous.
The memory actually allocated is stored in a new variable, passed to
internal functions as out parameter. It is then used to check whether
the repr uses less characters.