Print usage message to stdout
The usage text is not the display of an error, it is only written due
to the user asking for it, so it belongs in the standard output. Also,
the user might want to parse its contents (for example by piping to
grep) in which case it is easier when the text to parse is on stdout.