Remove restrict qualifiers for function params
The parameters to copy_bits and endswith are no longer qualified with
restrict. This is due to stop misusing it: copy_bits was actually used
to copy from overlapping areas (which restrict proscribes) and while
endswith did not yet, it very much could, in the future. So the keyword
is simply removed.