Pass a SDL_Color object to _draw_cell
The internal function had a high number of parameters, using an
SDL_Color object helps reduce this number by encapsulating four of them
into one. The object is passed directly by value because it is actually
smaller (or equal to in size, depending on the platform, but never
bigger) than a single pointer.