Add missing break at end of case block
This change concerns the case handling a W key press event.

Although it turns out that the break was not strictly necessary
(falling through to the nesxt cases had no side-effect until the next
break met) but it is still very confusing to read, and is most likely
an error.