Basically revert commit 67d5eb3
Moving the loop counter increment to the loop update statement had the
advantage of satisfying the linter, but the inconvenient of *not
working*! For some characters in the string (namely: new lines) the
variable must not be incremented. Indeed, it represents a bit index, or
a cell in the grid data. To carry this meaning more transparently, the
variable is renamed.