Merge plain text line ending handling blocks
In the internal function loading grid cells from a plain text repr, the
else-if branches for EOL chars (single LF, or CR followeed or not by
LF) are merged in a single else-if. Although the new condition
expression is more complex (it is the merging of three if expressions!)
the overall complexity of the block is reduced.
This change works nicely, because the error case of a CR not followed
by a LF is now encompassed by the last else-if expression, which is now
the only error exit point of the function.