Add parens around expr to shut up paranoid gcc
GCC complains when parens are *not* used around a logical AND
expression within a logical OR expression. It also does not support
-Wno-logical-op-parentheses, so the only way to silence the warning is
to oblige and add parentheses around the && expression.

This is plain stupid, and it makes *me* look like a newbie programmer.
I know perfectly well that && has higher precedence than ||, and *not*
using them here shows that I trust my programming skills, and that
other programmers can rely on *my* programming skills as well! This is
a tacit contract between developers that I am forced to break because
the GCC guys cater for newbies more than experienced programmers.
Anyway. At least the line stays below 80 chars, and I don't have to
wrap it.