Change toggle_cell to accept signed cordinates
This change is introduced to match the signature of get_grid_cell. The
implementation also mirrors this function's, delegating to either of
two new internal functions, depending on whether the grid wraps or not.
The performance impact will be minimal, since the function is only
called on user interaction.
The coordinates parameters are also renamed, once again to match
get_grid_cell.